pyproject.toml
[project]
name = "mcptest"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "fastmcp>=2.14.3",
    "jupyter>=1.1.1",
    "langchain>=1.2.3",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.7",
    "mcp[cli]>=1.25.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "yfinance>=1.0",
]
